requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-decouple==3.8
parameterized==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2022.1
openpyxl==3.1.2
drf-yasg==1.21.7